refactor instruction pod types into a separate submodule
