+++ zk-token-sdk/src/pod/mod.rs
@@ -1,10 +1,7 @@
-pub mod convert;
-// pub mod decryption;
-pub mod elgamal;
-// pub mod ops;
 pub mod auth_encryption;
+pub mod elgamal;
+pub mod instruction;
 pub mod pedersen;
-pub mod pod;
 pub mod range_proof;
 pub mod sigma_proofs;
 
@@ -18,6 +15,10 @@ pub use {
     auth_encryption::AeCiphertext,
     bytemuck::{Pod, Zeroable},
     elgamal::{DecryptHandle, ElGamalCiphertext, ElGamalPubkey},
+    instruction::{
+        FeeEncryption, FeeParameters, TransferAmountEncryption, TransferPubkeys,
+        TransferWithFeePubkeys,
+    },
     pedersen::PedersenCommitment,
     range_proof::{RangeProof128, RangeProof256, RangeProof64},
     sigma_proofs::{
